+++ NXNavigationExtension/Private/NXNavigationExtensionRuntime.h
@@ -141,7 +141,8 @@ NXNavigationExtensionExtendImplementationOfVoidMethodWithoutArguments(Class targ
 
 /// 将颜色转换成图片
 /// @param color UIColor
-CG_INLINE UIImage *NXNavigationExtensionGetImageFromColor(UIColor *color) {
+CG_INLINE UIImage *
+NXNavigationExtensionGetImageFromColor(UIColor *color) {
     CGSize size = CGSizeMake(1.0, 1.0);
     UIGraphicsBeginImageContextWithOptions(size, NO, 0);
     [color setFill];
@@ -153,7 +154,8 @@ CG_INLINE UIImage *NXNavigationExtensionGetImageFromColor(UIColor *color) {
 }
 
 /// 获取 Key window
-CG_INLINE UIWindow *NXNavigationExtensionGetKeyWindow(void) {
+CG_INLINE UIWindow *
+NXNavigationExtensionGetKeyWindow(void) {
     NSArray<UIWindow *> *windows = UIApplication.sharedApplication.windows;
     for (UIWindow *window in windows) {
         if (window.isKeyWindow) {